refactor: Migrate users.register endpoint to OpenAPI with AJV validation
