+++ .changeset/migrate-users-register-openapi.md
@@ -1,0 +1,7 @@
+---
+"@rocket.chat/meteor": minor
+"@rocket.chat/rest-typings": minor
+"@rocket.chat/web-ui-registration": minor
+---
+
+Add OpenAPI support for the `users.register` API endpoint by migrating to chained route definition with AJV body and response validation.
