+++ packages/web-ui-registration/src/users-register.d.ts
@@ -1,0 +1,19 @@
+import type { IUser } from '@rocket.chat/core-typings';
+
+declare module '@rocket.chat/rest-typings' {
+	interface Endpoints {
+		'/v1/users.register': {
+			POST: (params: {
+				username: string;
+				name?: string;
+				email: string;
+				pass: string;
+				secret?: string;
+				reason?: string;
+				customFields?: object;
+			}) => {
+				user: IUser;
+			};
+		};
+	}
+}
